fix pre render issue
