+++ src/app/auth/reset-password/page.tsx
@@ -1,150 +1,10 @@
-'use client'
-
-import { useState, useEffect } from 'react'
-import { useRouter } from 'next/navigation'
-import { Button } from '@/components/ui/button'
-
-import { Label } from '@/components/ui/label'
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { PasswordInput } from '@/components/ui/password-input'
-import { createClient } from '@/lib/supabase/client'
-
-// Force dynamic rendering to avoid build-time issues
+// Server Component wrapper for reset password page
 export const dynamic = 'force-dynamic'
-export const fetchCache = 'force-no-store'
+export const revalidate = 0
+export const runtime = 'nodejs'
+
+import ResetPasswordClient from '@/app/auth/reset-password/ResetPasswordClient'
 
 export default function ResetPasswordPage() {
-  const [isLoading, setIsLoading] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [success, setSuccess] = useState<string | null>(null)
-  const [password, setPassword] = useState('')
-  const [confirmPassword, setConfirmPassword] = useState('')
-  
-  const router = useRouter()
-
-  const supabase = createClient()
-
-  useEffect(() => {
-    // Check if we have the necessary tokens for password reset
-    const checkSession = async () => {
-      const { data: { session } } = await supabase.auth.getSession()
-      if (!session) {
-        setError('Invalid or expired reset link. Please request a new password reset.')
-      }
-    }
-    
-    checkSession()
-  }, [supabase.auth])
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsLoading(true)
-    setError(null)
-    setSuccess(null)
-
-    // Validate passwords
-    if (password.length < 8) {
-      setError('Password must be at least 8 characters long')
-      setIsLoading(false)
-      return
-    }
-
-    if (password !== confirmPassword) {
-      setError('Passwords do not match')
-      setIsLoading(false)
-      return
-    }
-
-    try {
-      const { error } = await supabase.auth.updateUser({
-        password: password
-      })
-
-      if (error) {
-        setError(error.message)
-      } else {
-        setSuccess('Password updated successfully! Redirecting to login...')
-        setTimeout(() => {
-          router.push('/login')
-        }, 2000)
-      }
-    } catch (error) {
-      setError('An unexpected error occurred')
-      console.error('Password update error:', error)
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 px-4">
-      <Card className="w-full max-w-md">
-        <CardHeader className="space-y-1">
-          <CardTitle className="text-2xl text-center">Set New Password</CardTitle>
-          <CardDescription className="text-center">
-            Enter your new password below
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <form onSubmit={handleSubmit} className="space-y-4">
-            <div className="space-y-2">
-              <Label htmlFor="password">New Password</Label>
-              <PasswordInput
-                id="password"
-                name="password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                placeholder="Enter new password"
-                required
-                disabled={isLoading}
-              />
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="confirmPassword">Confirm Password</Label>
-              <PasswordInput
-                id="confirmPassword"
-                name="confirmPassword"
-                value={confirmPassword}
-                onChange={(e) => setConfirmPassword(e.target.value)}
-                placeholder="Confirm new password"
-                required
-                disabled={isLoading}
-              />
-            </div>
-            
-            {error && (
-              <div className="text-sm text-red-600 bg-red-50 p-3 rounded-md">
-                {error}
-              </div>
-            )}
-            
-            {success && (
-              <div className="text-sm text-green-600 bg-green-50 p-3 rounded-md">
-                {success}
-              </div>
-            )}
-            
-            <Button 
-              type="submit" 
-              className="w-full" 
-              disabled={isLoading}
-            >
-              {isLoading ? 'Updating...' : 'Update Password'}
-            </Button>
-          </form>
-          
-          <div className="mt-4 text-center text-sm">
-            <Button 
-              variant="link" 
-              onClick={() => router.push('/login')}
-              className="p-0 h-auto"
-            >
-              Back to Sign In
-            </Button>
-          </div>
-        </CardContent>
-      </Card>
-    </div>
-  )
+  return <ResetPasswordClient />
 }
